# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Microblogging API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /nav-links:
    get:
      operationId: listNavLinks
      tags: [nav-links]
      summary: List the owner-configured navbar links
      description: |
        Public read. Returns every visible nav link sorted ascending by
        `sortOrder`. Owners may pass `includeHidden=1` to also receive
        rows with `visible=false` (used by /admin/navigation and
        /admin/pages so the owner can manage hidden rows).
      parameters:
        - in: query
          name: includeHidden
          required: false
          schema:
            type: string
            enum: ["1"]
          description: When `1` and the caller is the owner, includes hidden rows.
      responses:
        "200":
          description: Nav-links list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NavLinksList"
    post:
      operationId: createNavLink
      tags: [nav-links]
      summary: Create a new nav link (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNavLinkBody"
      responses:
        "201":
          description: Created nav link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NavLink"
        "400":
          description: Invalid body
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /nav-links/{id}:
    patch:
      operationId: updateNavLink
      tags: [nav-links]
      summary: Update a nav link (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNavLinkBody"
      responses:
        "200":
          description: Updated nav link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NavLink"
        "400":
          description: Invalid body
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteNavLink
      tags: [nav-links]
      summary: Delete a nav link (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /nav-items/reorder:
    patch:
      operationId: reorderNavItems
      tags: [nav-items]
      summary: Persist a new order for nav items (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NavItemsReorderBody"
      responses:
        "200":
          description: Reordered nav items list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NavLinksList"
        "400":
          description: Invalid body
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
components:
  schemas:
...
    NavLink:
      type: object
      description: |
        A single navbar entry. The `kind` discriminator drives rendering:

          * `external` — `url` is an absolute external URL (Task #24).
          * `page`     — `pageId` and `pageSlug` are set; the public href
                         is `/p/<pageSlug>` resolved server-side via JOIN.
          * `system`   — built-in route (e.g. `/feeds`); the owner can
                         hide via `visible=false` but cannot delete.
      properties:
        id:
          type: integer
        label:
          type: string
          maxLength: 64
        url:
          type: string
          maxLength: 2048
        openInNewTab:
          type: boolean
        sortOrder:
          type: integer
        kind:
          type: string
          enum: [external, page, system]
        pageId:
          type: integer
          nullable: true
        pageSlug:
          type: string
          nullable: true
          description: Resolved at read time via JOIN on pages.id when kind='page'.
        visible:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - label
        - url
        - openInNewTab
        - sortOrder
        - kind
        - visible
        - createdAt
        - updatedAt
    NavLinksList:
      type: object
      properties:
        links:
          type: array
          items:
            $ref: "#/components/schemas/NavLink"
      required:
        - links
    CreateNavLinkBody:
      type: object
      properties:
        label:
          type: string
          maxLength: 64
        url:
          type: string
          maxLength: 2048
        openInNewTab:
          type: boolean
          default: true
        sortOrder:
          type: integer
          default: 0
      required:
        - label
        - url
    UpdateNavLinkBody:
      type: object
      properties:
        label:
          type: string
          maxLength: 64
        url:
          type: string
          maxLength: 2048
        openInNewTab:
          type: boolean
        sortOrder:
          type: integer
        visible:
          type: boolean
    NavItemsReorderBody:
      type: object
      description: |
        Persist a new order for every nav item. The `items` array MUST
        include every existing nav row exactly once (the server rejects
        partial reorders with `400` and lists the missing/unknown ids).
        The server applies the updates in a single transaction and
        renumbers in increments of 10 to keep gaps tidy. Clients
        typically send `sortOrder = (i+1)*10` for each item; the server
        also accepts arbitrary positive ordinals and renumbers based on
        their relative order.
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              sortOrder:
                type: integer
            required:
              - id
              - sortOrder
      required:
        - items
...